fix(rules): define an explicit error on non-concurrent index drops
