+++ linter/src/rules/mod.rs
@@ -25,7 +25,9 @@ pub use ban_drop_column::*;
 pub mod require_concurrent_index_creation;
 pub use require_concurrent_index_creation::*;
 pub mod ban_char_field;
+pub mod require_concurrent_index_deletion;
 pub use ban_char_field::*;
+pub use require_concurrent_index_deletion::*;
 pub mod adding_foreign_key_constraint;
 pub use adding_foreign_key_constraint::*;
 mod utils;
