+++ linter/src/rules/require_concurrent_index_creation.rs
@@ -1,6 +1,6 @@
 use crate::rules::utils::tables_created_in_transaction;
 use crate::violations::{RuleViolation, RuleViolationKind};
-use squawk_parser::ast::{ObjectType, RawStmt, Stmt};
+use squawk_parser::ast::{RawStmt, Stmt};
 
 #[must_use]
 pub fn require_concurrent_index_creation(tree: &[RawStmt]) -> Vec<RuleViolation> {
@@ -19,13 +19,6 @@ pub fn require_concurrent_index_creation(tree: &[RawStmt]) -> Vec<RuleViolation>
                     ));
                 }
             }
-            Stmt::DropStmt(stmt) if !stmt.concurrent && stmt.remove_type == ObjectType::Index => {
-                errs.push(RuleViolation::new(
-                    RuleViolationKind::RequireConcurrentIndexCreation,
-                    raw_stmt.into(),
-                    None,
-                ))
-            }
             _ => continue,
         }
     }
@@ -35,7 +28,6 @@ pub fn require_concurrent_index_creation(tree: &[RawStmt]) -> Vec<RuleViolation>
 #[cfg(test)]
 mod test_rules {
     use crate::check_sql;
-    use crate::violations::RuleViolationKind;
     use insta::assert_debug_snapshot;
 
     /// ```sql
@@ -58,48 +50,5 @@ mod test_rules {
   CREATE INDEX CONCURRENTLY "field_name_idx" ON "table_name" ("field_name");
   "#;
         assert_debug_snapshot!(check_sql(ok_sql, &["prefer-robust-stmts".into()]));
-    }
-
-    #[test]
-    fn test_drop_index_concurrently() {
-        let bad_sql = r#"
-  -- instead of
-  DROP INDEX IF EXISTS "field_name_idx";
-  "#;
-        let res = check_sql(bad_sql, &[]).unwrap();
-        assert_eq!(res.len(), 1);
-        assert_eq!(
-            res[0].kind,
-            RuleViolationKind::RequireConcurrentIndexCreation
-        );
-
-        let ok_sql = r#"
-  DROP INDEX CONCURRENTLY IF EXISTS "field_name_idx";
-  "#;
-        assert_eq!(check_sql(ok_sql, &[]), Ok(vec![]));
-    }
-
-    #[test]
-    fn regression_false_positive_drop_type() {
-        let sql = r#"
-  DROP TYPE IF EXISTS foo;
-  "#;
-        assert_eq!(check_sql(sql, &[]), Ok(vec![]));
-    }
-
-    #[test]
-    fn regression_false_positive_drop_table() {
-        let sql = r#"
-  DROP TABLE IF EXISTS some_table;
-  "#;
-        assert_eq!(check_sql(sql, &[]), Ok(vec![]));
-    }
-
-    #[test]
-    fn regression_false_positive_drop_trigger() {
-        let sql = r#"
-  DROP TRIGGER IF EXISTS trigger on foo_table;
-  "#;
-        assert_eq!(check_sql(sql, &[]), Ok(vec![]));
     }
 }
